Refactor: 코드리뷰 반영 refactor - scss 파일 공통 사용 부분 함수화 - utils 하위에 함수마다 분리하여 사용
