+++ src/utils/printNumberWithComma.js
@@ -1,0 +1,2 @@
+export const printNumberWithComma = number =>
+  number.toString().replace(/\B(?=(\d{3})+(?!\d))/g, ',');
